+++ UISystem/FramerateUI.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "imgui.h"
+#include "UIElementsGroup.h"
 
 namespace NCL {
 	namespace UI {
@@ -9,7 +10,9 @@ namespace NCL {
 			FramerateUI();
 			~FramerateUI();
 
-			void Draw(float dt, int winWidth, int winHeight);
+			CSC8508::PushdownState::PushdownResult OnElementsRender(float dt);
+
+			/*void Draw(float dt, int winWidth, int winHeight);*/
 
 		protected:
 			float dt = 0;
